Add the following optimization
Add parameter names in the header file.
Remove IsEmpty() from .cpp and define it in the header.
Remove m_NoOfElems, and m_NoOfNodes from destructor (This is pointless)
Remove manual swap in MirrorTree() and use std::swap() for that.
Shorten = operator overload function.
Instead of defining all the variables on top of the function, I change it
to define things when they are about to use.